app/requirements.txt
flask
mysql-connector-python
redis
celery